# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DiscordStatus {
  configured: boolean;
  connected: boolean;
  /** @nullable */
  botUser: string | null;
  message: string;
}

export interface Guild {
  id: string;
  name: string;
  /** @nullable */
  iconUrl: string | null;
  /** @minimum 0 */
  memberCount: number;
  canManage: boolean;
  botPresent: boolean;
  /** @nullable */
  inviteUrl: string | null;
}

export interface DiscordChannel {
  id: string;
  name: string;
  type: string;
}

export interface DiscordRole {
  id: string;
  name: string;
  color: string;
  /** @minimum 0 */
  position: number;
  managed: boolean;
}

export type GuildDetails = Guild & {
  channels: DiscordChannel[];
  roles: DiscordRole[];
};

export interface ActivityItem {
  id: string;
  kind: string;
  title: string;
  detail: string;
  createdAt: string;
}

export interface DashboardSummary {
  /** @minimum 0 */
  guildCount: number;
  /** @minimum 0 */
  configuredGuildCount: number;
  /** @minimum 0 */
  totalMembers: number;
  /** @minimum 0 */
  recentJoins: number;
  inviteAttributionRate: number;
  latestActivity?: ActivityItem[];
}

export type WelcomeSettingsStyle = typeof WelcomeSettingsStyle[keyof typeof WelcomeSettingsStyle];


export const WelcomeSettingsStyle = {
  embed: 'embed',
  banner: 'banner',
} as const;

export type CardDesignBackgroundMode = typeof CardDesignBackgroundMode[keyof typeof CardDesignBackgroundMode];


export const CardDesignBackgroundMode = {
  url: 'url',
  transparent: 'transparent',
} as const;

export type CardDesignAvatarShape = typeof CardDesignAvatarShape[keyof typeof CardDesignAvatarShape];


export const CardDesignAvatarShape = {
  square: 'square',
  rounded: 'rounded',
  circle: 'circle',
} as const;

export type CardDesignUsernameFont = typeof CardDesignUsernameFont[keyof typeof CardDesignUsernameFont];


export const CardDesignUsernameFont = {
  CairoBold: 'CairoBold',
  ChangaBold: 'ChangaBold',
  AlmaraiBold: 'AlmaraiBold',
} as const;

export type CardDesignUsernameAlign = typeof CardDesignUsernameAlign[keyof typeof CardDesignUsernameAlign];


export const CardDesignUsernameAlign = {
  right: 'right',
  center: 'center',
  left: 'left',
} as const;

export type CardDesignAvatar = {
  enabled: boolean;
  x: number;
  y: number;
  /**
     * @minimum 32
     * @maximum 600
     */
  size: number;
  shape: CardDesignAvatarShape;
};

export type CardDesignUsername = {
  enabled: boolean;
  /** @maxLength 500 */
  text: string;
  x: number;
  y: number;
  font: CardDesignUsernameFont;
  /**
     * @minimum 8
     * @maximum 160
     */
  size: number;
  /**
     * @minimum 0.8
     * @maximum 3
     */
  lineHeight: number;
  align: CardDesignUsernameAlign;
  /** @pattern ^#[0-9a-fA-F]{6}$ */
  color: string;
  shadow: boolean;
  /** @pattern ^#[0-9a-fA-F]{6}$ */
  shadowColor: string;
  /**
     * @minimum -50
     * @maximum 50
     */
  shadowX: number;
  /**
     * @minimum -50
     * @maximum 50
     */
  shadowY: number;
  /**
     * @minimum 0
     * @maximum 60
     */
  shadowBlur: number;
};

export type CardDesignExtra = {
  enabled: boolean;
  /** @nullable */
  url: string | null;
  x: number;
  y: number;
  /**
     * @minimum 1
     * @maximum 1200
     */
  width: number;
  /**
     * @minimum 1
     * @maximum 1200
     */
  height: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  opacity: number;
};

export interface CardDesign {
  /**
     * @minimum 320
     * @maximum 2400
     */
  width: number;
  /**
     * @minimum 180
     * @maximum 1400
     */
  height: number;
  backgroundMode: CardDesignBackgroundMode;
  /** @nullable */
  backgroundUrl: string | null;
  avatar: CardDesignAvatar;
  username: CardDesignUsername;
  extra: CardDesignExtra;
}

export interface MessageSuite {
  welcomeChannelEnabled: boolean;
  /** @maxLength 1024 */
  welcomeMessage: string;
  dmEnabled: boolean;
  /** @maxLength 1024 */
  dmMessage: string;
  leaveEnabled: boolean;
  /** @maxLength 1024 */
  leaveMessage: string;
}

export interface CommandConfig {
  name: string;
  enabled: boolean;
  aliases: string[];
  allowedRoleIds: string[];
  blockedChannelIds: string[];
  deleteCommand: boolean;
  deleteBotResponse: boolean;
}

export interface WelcomeSettings {
  guildId: string;
  enabled: boolean;
  style: WelcomeSettingsStyle;
  /** @nullable */
  channelId: string | null;
  headline: string;
  body: string;
  accentColor: string;
  /** @nullable */
  backgroundUrl: string | null;
  includeInviter: boolean;
  autoRoleIds: string[];
  cardDesign?: CardDesign;
  messageSuite?: MessageSuite;
  commandConfig?: CommandConfig[];
  /** @nullable */
  updatedAt: string | null;
}

export type WelcomeSettingsInputStyle = typeof WelcomeSettingsInputStyle[keyof typeof WelcomeSettingsInputStyle];


export const WelcomeSettingsInputStyle = {
  embed: 'embed',
  banner: 'banner',
} as const;

export interface WelcomeSettingsInput {
  enabled: boolean;
  style: WelcomeSettingsInputStyle;
  /** @nullable */
  channelId: string | null;
  headline: string;
  body: string;
  accentColor: string;
  /** @nullable */
  backgroundUrl: string | null;
  includeInviter: boolean;
  autoRoleIds: string[];
  cardDesign?: CardDesign;
  messageSuite?: MessageSuite;
  commandConfig?: CommandConfig[];
}