# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityItem {
  id: string;
  kind: string;
  title: string;
  detail: string;
  createdAt: Date;
}