# lib/api-zod/src/generated/types/cardDesign.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CardDesignAvatar } from './cardDesignAvatar';
import type { CardDesignBackgroundMode } from './cardDesignBackgroundMode';
import type { CardDesignExtra } from './cardDesignExtra';
import type { CardDesignUsername } from './cardDesignUsername';

export interface CardDesign {
  /**
     * @minimum 320
     * @maximum 2400
     */
  width: number;
  /**
     * @minimum 180
     * @maximum 1400
     */
  height: number;
  backgroundMode: CardDesignBackgroundMode;
  /** @nullable */
  backgroundUrl: string | null;
  avatar: CardDesignAvatar;
  username: CardDesignUsername;
  extra: CardDesignExtra;
}