# lib/api-zod/src/generated/types/commandConfig.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CommandConfig {
  name: string;
  enabled: boolean;
  aliases: string[];
  allowedRoleIds: string[];
  blockedChannelIds: string[];
  deleteCommand: boolean;
  deleteBotResponse: boolean;
}