# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItem } from './activityItem';

export interface DashboardSummary {
  /** @minimum 0 */
  guildCount: number;
  /** @minimum 0 */
  configuredGuildCount: number;
  /** @minimum 0 */
  totalMembers: number;
  /** @minimum 0 */
  recentJoins: number;
  inviteAttributionRate: number;
  latestActivity?: ActivityItem[];
}