# lib/api-zod/src/generated/types/discordRole.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DiscordRole {
  id: string;
  name: string;
  color: string;
  /** @minimum 0 */
  position: number;
  managed: boolean;
}