# lib/api-zod/src/generated/types/guildDetails.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DiscordChannel } from './discordChannel';
import type { DiscordRole } from './discordRole';
import type { Guild } from './guild';

export type GuildDetails = Guild & {
  channels: DiscordChannel[];
  roles: DiscordRole[];
};