# lib/api-zod/src/generated/types/messageSuite.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MessageSuite {
  welcomeChannelEnabled: boolean;
  /** @maxLength 1024 */
  welcomeMessage: string;
  dmEnabled: boolean;
  /** @maxLength 1024 */
  dmMessage: string;
  leaveEnabled: boolean;
  /** @maxLength 1024 */
  leaveMessage: string;
}