# lib/api-zod/src/generated/types/welcomeSettings.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CardDesign } from './cardDesign';
import type { CommandConfig } from './commandConfig';
import type { MessageSuite } from './messageSuite';
import type { WelcomeSettingsStyle } from './welcomeSettingsStyle';

export interface WelcomeSettings {
  guildId: string;
  enabled: boolean;
  style: WelcomeSettingsStyle;
  /** @nullable */
  channelId: string | null;
  headline: string;
  body: string;
  accentColor: string;
  /** @nullable */
  backgroundUrl: string | null;
  includeInviter: boolean;
  autoRoleIds: string[];
  cardDesign?: CardDesign;
  messageSuite?: MessageSuite;
  commandConfig?: CommandConfig[];
  /** @nullable */
  updatedAt: Date | null;
}