# lib/api-zod/src/generated/types/welcomeSettingsInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CardDesign } from './cardDesign';
import type { CommandConfig } from './commandConfig';
import type { MessageSuite } from './messageSuite';
import type { WelcomeSettingsInputStyle } from './welcomeSettingsInputStyle';

export interface WelcomeSettingsInput {
  enabled: boolean;
  style: WelcomeSettingsInputStyle;
  /** @nullable */
  channelId: string | null;
  headline: string;
  body: string;
  accentColor: string;
  /** @nullable */
  backgroundUrl: string | null;
  includeInviter: boolean;
  autoRoleIds: string[];
  cardDesign?: CardDesign;
  messageSuite?: MessageSuite;
  commandConfig?: CommandConfig[];
}